requirements.txt
requests==2.22.0
statsmodels==0.9.0
flask== 1.0.2
pyaml==18.11.0
python-dateutil==2.8.0
SQLAlchemy==1.3.1
flask_sqlalchemy==2.3.2
pytest==4.0.1
boto3==1.9.145
botocore==1.12.164
scipy==1.2.1
PyYAML==5.1.1
PyMySQL==0.9.3
pandas==0.24.2
numpy==1.16.4